feat: delete button deletes both docs and files related to qr codes

+++ src/myHooks/deleteDocFunction.js
@@ -1,7 +1,9 @@
 import { deleteDoc, doc } from "firebase/firestore";
-import { db, auth } from "../../config/firebase";
+import { db, auth, storage } from "../../config/firebase";
+// import { storage } from "../../config/firebase";
+import { deleteObject, ref } from "firebase/storage";
 
-export const deleteDocFunction = async (id) => {
+export const deleteDocFunction = async (id, type, fileName) => {
   const user = auth.currentUser.email;
   const collectionRef = doc(
     db,
@@ -10,5 +12,22 @@ export const deleteDocFunction = async (id) => {
     "qr-code-data",
     id
   );
-  await deleteDoc(collectionRef);
+  if (type === "file") {
+    const bucketName = storage.app.options.storageBucket;
+    const gsPath = `gs://${bucketName}/files/${fileName}`;
+
+    const fileRef = ref(storage, gsPath);
+    try {
+      const isDeleted = deleteObject(fileRef);
+      if (isDeleted) {
+        alert("deleted");
+        console.log(isDeleted);
+        await deleteDoc(collectionRef);
+      }
+    } catch (error) {
+      console.log(error);
+    }
+  } else {
+    await deleteDoc(collectionRef);
+  }
 };
